requirements.txt
Flask~=1.1.2
pyldap~=3.0.0.post1
PyYAML~=5.4.1